Fixed but where MSTC would be less than active + completed, causing a miscalc
This issue goes deeper than the 3cpm. It seems to be directly in 3C. You can set MSTC to 0 while still having an active SO on the exchange. What this means is in the calc before the active SO was ignored completely. This now will find the highest and use that as the total SOs. It's not a perfect fix, but in testing this made the mstc correct and the metrics matched.

github issue - https://github.com/3commas-io/3commas-official-api-docs/issues/93

This may need further work to build in manual SOs.