Updated the manual SO count metric - #29
This metric coming from 3C can have the wrong data coming from it for `completed_manual_safety_orders`. This leads to a skew in the metric view on Active deals themselves. Went the route of updating the same value with the database. Long term when redoing the api calls this will need to be looked at.