V0.4.1 (#32)
* Style updates for the active deals table

* Fixed label wrapping on bot perf charts

* Fixed hover on darkmode button

* Initial try of the signed mac os app plus fixed with active deal numbers

* Fixed but where MSTC would be less than active + completed, causing a miscalc

This issue goes deeper than the 3cpm. It seems to be directly in 3C. You can set MSTC to 0 while still having an active SO on the exchange. What this means is in the calc before the active SO was ignored completely. This now will find the highest and use that as the total SOs. It's not a perfect fix, but in testing this made the mstc correct and the metrics matched.

github issue - https://github.com/3commas-io/3commas-official-api-docs/issues/93

This may need further work to build in manual SOs.

* Updated the manual SO count metric - #29

This metric coming from 3C can have the wrong data coming from it for `completed_manual_safety_orders`. This leads to a skew in the metric view on Active deals themselves. Went the route of updating the same value with the database. Long term when redoing the api calls this will need to be looked at.

* Adding a custom menu bar with links.

* Updating readme with FAQ, API docs, and the changelog

* Fixed inactive funds calc for bots

the way 3C returns the data it's possible for max_active_deals to be lower than active_deals_count. This causes an negative value to be introduced and skew the metrics.

* Fixed bug in card causing an findDOMNode error with the day profit

* Updated changelog

* add key to the coin listing to remove react warning (#30)

Co-authored-by: Julien Tant <julien@craftyx.fr>